# openApi/openapi.yaml
openapi: 3.0.0
info:
  title: Parking Lot Service API
  version: 1.0.0
paths:
  /parking-lots:
    get:
      summary: Get all parking lots
      operationId: getParkingLots
      responses:
        '200':
          description: A list of parking lots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParkingLot'
    post:
      summary: Create a new parking lot
      operationId: createParkingLot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingLot'
      responses:
        '201':
          description: The created parking lot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingLot'
  /parking-lots/{id}:
    get:
      summary: Get a parking lot by ID
      operationId: getParkingLotById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: The parking lot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingLot'
    put:
      summary: Update a parking lot by ID
      operationId: updateParkingLotById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingLot'
      responses:
        '200':
          description: The updated parking lot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingLot'
    delete:
      summary: Delete a parking lot by ID
      operationId: deleteParkingLotById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: The parking lot was deleted

components:
  schemas:
    ParkingLot:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        spots:
          type: object
          properties:
            motorcycles:
              type: integer
            cars:
              type: integer
            buses:
              type: integer
    EntryTicket:
      type: object
      properties:
        id:
          type: integer
        vehicleType:
          type: string
        entryTime:
          type: string
          format: date-time
        parkingLot:
          $ref: '#/components/schemas/ParkingLot'
    Receipt:
      type: object
      properties:
        id:
          type: integer
        vehicleType:
          type: string
        exitTime:
          type: string
          format: date-time
        fee:
          type: number
          format: float
        parkingLot:
          $ref: '#/components/schemas/ParkingLot'